✨ F08: ORDER > MENU > Add resetMenu and create EmptyMenuClient and EmptyMenuAdmin
