+++ src/components/pages/order/Main/Menu/Menu.jsx
@@ -1,0 +1,52 @@
+import { useContext } from "react";
+import styled from "styled-components";
+import OrderContext from "../../../../../context/OrderContext";
+import { theme } from "../../../../../theme";
+import { formatPrice } from "../../../../../utils/maths";
+import Card from "../../../../reusable-ui/Card";
+import EmptyMenuAdmin from "./EmptyMenuAdmin";
+import EmptyMenuClient from "./EmptyMenuClient";
+
+export default function Menu() {
+	const { menu, isModeAdmin, handleDelete, resetMenu } = useContext(OrderContext);
+	const IMAGE_BY_DEFAULT = "../../../../../public/images/coming-soon.png";
+
+	if (menu.length == 0) {
+		return isModeAdmin ? (
+			<EmptyMenuAdmin resetMenu={resetMenu} classname="page-without-product" />
+		) : (
+			<EmptyMenuClient classname="page-without-product" />
+		);
+	}
+
+	return (
+		<MenuStyled>
+			{menu.map(({ id, title, imageSource, price }) => {
+				return (
+					<Card
+						key={id}
+						imageSource={imageSource ? imageSource : IMAGE_BY_DEFAULT}
+						title={title}
+						leftDescription={formatPrice(price)}
+						hasDeleteButton={isModeAdmin}
+						onClick={() => handleDelete(id)}
+					/>
+				);
+			})}
+		</MenuStyled>
+	);
+}
+
+const MenuStyled = styled.div`
+	background: ${theme.colors.background_white};
+	box-shadow: ${theme.shadows.strong};
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+	grid-row-gap: 60px;
+	padding: 50px 50px 150px;
+	justify-items: center;
+	overflow-y: scroll;
+	&::-webkit-scrollbar {
+		display: none;
+	}
+`;
