+++ src/components/reusable-ui/PrimaryButton.jsx
@@ -1,9 +1,9 @@
 import styled from "styled-components";
 import { theme } from "../../theme";
 
-export default function PrimaryButton({ label, Icon, className }) {
+export default function PrimaryButton({ label, Icon, className, onClick }) {
 	return (
-		<PrimaryButtonStyled className={className}>
+		<PrimaryButtonStyled className={className} onClick={onClick}>
 			<span>{label}</span>
 			{Icon && Icon}
 		</PrimaryButtonStyled>
