+++ src/context/OrderContext.jsx
@@ -15,4 +15,6 @@ export default createContext({
 
 	handleAdd: () => {},
 	handleDelete: () => {},
+
+	resetMenu: () => {},
 });
